LayoutTests/imported/w3c/web-platform-tests/wasm/core/unreached-invalid.wast.js-expected.txt


PASS #1 Reinitialize the default imports
PASS #2 Test that WebAssembly compilation fails
PASS #3 Test that WebAssembly compilation fails
PASS #4 Test that WebAssembly compilation fails
PASS #5 Test that WebAssembly compilation fails
FAIL #6 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #7 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #8 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #9 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #10 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #11 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #12 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #13 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #14 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #15 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #16 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #17 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #18 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #19 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #20 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #21 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #22 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #23 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #24 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #25 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #26 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #27 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #28 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #29 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #30 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #31 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #32 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #33 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #34 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #35 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #36 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #37 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #38 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #39 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #40 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #41 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #42 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #43 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #44 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #45 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #46 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #47 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #48 Test that WebAssembly compilation fails assert_equals: expected true but got false
PASS #49 Test that WebAssembly compilation fails
PASS #50 Test that WebAssembly compilation fails
PASS #51 Test that WebAssembly compilation fails
PASS #52 Test that WebAssembly compilation fails
PASS #53 Test that WebAssembly compilation fails
PASS #54 Test that WebAssembly compilation fails
PASS #55 Test that WebAssembly compilation fails
PASS #56 Test that WebAssembly compilation fails
PASS #57 Test that WebAssembly compilation fails
PASS #58 Test that WebAssembly compilation fails
PASS #59 Test that WebAssembly compilation fails
PASS #60 Test that WebAssembly compilation fails
PASS #61 Test that WebAssembly compilation fails
PASS #62 Test that WebAssembly compilation fails
PASS #63 Test that WebAssembly compilation fails
PASS #64 Test that WebAssembly compilation fails
PASS #65 Test that WebAssembly compilation fails
PASS #66 Test that WebAssembly compilation fails
PASS #67 Test that WebAssembly compilation fails
PASS #68 Test that WebAssembly compilation fails
PASS #69 Test that WebAssembly compilation fails
PASS #70 Test that WebAssembly compilation fails
PASS #71 Test that WebAssembly compilation fails
PASS #72 Test that WebAssembly compilation fails
PASS #73 Test that WebAssembly compilation fails
PASS #74 Test that WebAssembly compilation fails
PASS #75 Test that WebAssembly compilation fails
PASS #76 Test that WebAssembly compilation fails
PASS #77 Test that WebAssembly compilation fails
PASS #78 Test that WebAssembly compilation fails
PASS #79 Test that WebAssembly compilation fails
FAIL #80 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #81 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #82 Test that WebAssembly compilation fails assert_equals: expected true but got false
PASS #83 Test that WebAssembly compilation fails
FAIL #84 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #85 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #86 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #87 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #88 Test that WebAssembly compilation fails assert_equals: expected true but got false
PASS #89 Test that WebAssembly compilation fails
PASS #90 Test that WebAssembly compilation fails
PASS #91 Test that WebAssembly compilation fails
PASS #92 Test that WebAssembly compilation fails
PASS #93 Test that WebAssembly compilation fails
PASS #94 Test that WebAssembly compilation fails
PASS #95 Test that WebAssembly compilation fails
PASS #96 Test that WebAssembly compilation fails
PASS #97 Test that WebAssembly compilation fails
PASS #98 Test that WebAssembly compilation fails
PASS #99 Test that WebAssembly compilation fails
PASS #100 Test that WebAssembly compilation fails
PASS #101 Test that WebAssembly compilation fails
PASS #102 Test that WebAssembly compilation fails
PASS #103 Test that WebAssembly compilation fails
PASS #104 Test that WebAssembly compilation fails
PASS #105 Test that WebAssembly compilation fails
PASS #106 Test that WebAssembly compilation fails
PASS #107 Test that WebAssembly compilation fails
PASS #108 Test that WebAssembly compilation fails
FAIL #109 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #110 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #111 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #112 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #113 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #114 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #115 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #116 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #117 Test that WebAssembly compilation fails assert_equals: expected true but got false
FAIL #118 Test that WebAssembly compilation fails assert_equals: expected true but got false
PASS #119 Test that WebAssembly compilation fails
PASS #120 Reinitialize the default imports
PASS #121 Test that WebAssembly compilation fails
PASS #122 Test that WebAssembly compilation fails
PASS #123 Test that WebAssembly compilation fails
PASS #124 Test that WebAssembly compilation fails
FAIL #125 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:16:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #126 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:19:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #127 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:22:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #128 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:25:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #129 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:28:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #130 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:31:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #131 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:34:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #132 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:37:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #133 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:40:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #134 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:43:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #135 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:46:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #136 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:49:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #137 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:52:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #138 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:55:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #139 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:58:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #140 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:61:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #141 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:64:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #142 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:67:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #143 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:70:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #144 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:73:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #145 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:76:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #146 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:79:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #147 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:82:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #148 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:85:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #149 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:88:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #150 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:91:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #151 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:94:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #152 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:97:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #153 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:100:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #154 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:103:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #155 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:106:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #156 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:109:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #157 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:112:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #158 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:115:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #159 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:118:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #160 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:121:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #161 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:124:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #162 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:127:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #163 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:130:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #164 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:133:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #165 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:136:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #166 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:139:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #167 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:142:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
PASS #168 Test that WebAssembly compilation fails
PASS #169 Test that WebAssembly compilation fails
PASS #170 Test that WebAssembly compilation fails
PASS #171 Test that WebAssembly compilation fails
PASS #172 Test that WebAssembly compilation fails
PASS #173 Test that WebAssembly compilation fails
PASS #174 Test that WebAssembly compilation fails
PASS #175 Test that WebAssembly compilation fails
PASS #176 Test that WebAssembly compilation fails
PASS #177 Test that WebAssembly compilation fails
PASS #178 Test that WebAssembly compilation fails
PASS #179 Test that WebAssembly compilation fails
PASS #180 Test that WebAssembly compilation fails
PASS #181 Test that WebAssembly compilation fails
PASS #182 Test that WebAssembly compilation fails
PASS #183 Test that WebAssembly compilation fails
PASS #184 Test that WebAssembly compilation fails
PASS #185 Test that WebAssembly compilation fails
PASS #186 Test that WebAssembly compilation fails
PASS #187 Test that WebAssembly compilation fails
PASS #188 Test that WebAssembly compilation fails
PASS #189 Test that WebAssembly compilation fails
PASS #190 Test that WebAssembly compilation fails
PASS #191 Test that WebAssembly compilation fails
PASS #192 Test that WebAssembly compilation fails
PASS #193 Test that WebAssembly compilation fails
PASS #194 Test that WebAssembly compilation fails
PASS #195 Test that WebAssembly compilation fails
PASS #196 Test that WebAssembly compilation fails
PASS #197 Test that WebAssembly compilation fails
PASS #198 Test that WebAssembly compilation fails
FAIL #199 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:238:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #200 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:241:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #201 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:244:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
PASS #202 Test that WebAssembly compilation fails
FAIL #203 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:250:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #204 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:253:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #205 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:256:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #206 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:259:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #207 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:262:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
PASS #208 Test that WebAssembly compilation fails
PASS #209 Test that WebAssembly compilation fails
PASS #210 Test that WebAssembly compilation fails
PASS #211 Test that WebAssembly compilation fails
PASS #212 Test that WebAssembly compilation fails
PASS #213 Test that WebAssembly compilation fails
PASS #214 Test that WebAssembly compilation fails
PASS #215 Test that WebAssembly compilation fails
PASS #216 Test that WebAssembly compilation fails
PASS #217 Test that WebAssembly compilation fails
PASS #218 Test that WebAssembly compilation fails
PASS #219 Test that WebAssembly compilation fails
PASS #220 Test that WebAssembly compilation fails
PASS #221 Test that WebAssembly compilation fails
PASS #222 Test that WebAssembly compilation fails
PASS #223 Test that WebAssembly compilation fails
PASS #224 Test that WebAssembly compilation fails
PASS #225 Test that WebAssembly compilation fails
PASS #226 Test that WebAssembly compilation fails
PASS #227 Test that WebAssembly compilation fails
FAIL #228 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:325:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #229 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:328:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #230 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:331:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #231 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:334:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #232 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:337:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #233 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:340:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #234 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:343:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #235 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:346:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #236 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:349:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
FAIL #237 Test that WebAssembly compilation fails assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
assert_invalid@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:174:9
unreached_invalid_wast_js@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:352:15
global code@http://web-platform.test:8800/wasm/core/js/unreached-invalid.wast.js:357:3 expected true got false
PASS #238 Test that WebAssembly compilation fails
